misc: ast_vbios: Add DAC control
Add control packer for DAC usage in this version.
In the current condition, the DAC resoultion should
be smaller than 1280 x 1024 when the VLIK was set
788 MHz. Beside the resolution changed could not from
smaller into larger mode. It would hit the packer isuue
and let the VLINK hang. When the EDID function was
turned on then the monitor fitting mode should be
smaller then 1280 x 1024.

Signed-off-by: Tommy Huang <tommy_huang@aspeedtech.com>
Change-Id: Ie53a53017a16199169aae8490871bd60d5cbad75