net: phy: ncsi: fixed not nullify the pointers after free
When NC-SI driver receives the ANE (Asynchronous Event Notification)
packets, it redo the process of NC-SI protocol.
Since the pointers not nullify after free and the driver continues to
use them, the pointers will reallocate at probe stage.
It happens the UAF (use-after-free) to cause the double free when do
the reallo function for the pointers.
The situation will cause the memory management occurs error.

Signed-off-by: Jacky Chou <jacky_chou@aspeedtech.com>
Change-Id: I1cf2a868e06e8f854b5d6b1714e580903c684ffd