cmd: ltpi-v2: Remove modification of hardware registers
The hardware strap result is stored in the SCU register. The original
driver modifies this register and initializes the LTPI interface based on
its value. However, this approach causes unintended behavior after a BMC
reboot, because the ROM code also relies on the SCU register. Since the
hardware strap is not reloaded into the SCU after reset, modifying it
leads to inconsistencies.
This patch resolves the issue by using a software variable to store the
configuration passed through the LTPI command, instead of altering the SCU
register directly.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I05ea0668d52acf4df7e16ca26e1a187fdffce1cd