ltpi: add SGPIO_PARALLEL_OUT_PROTECT for A1 compatibility
Adds the SGPIO_PARALLEL_OUT_PROTECT bit to the configuration to ensure
proper behavior on A1 hardware. Notes that this setting should be
removed when migrating to A2 due to an inversion in the bit's meaning.
This bit is used to protect the parallel output to be csr before serial
input is valid.

Change-Id: I2bfde390eb44977ed943adba02ca5140e32fd29b
Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>