ltpi: add support for serial output source selection
Introduce functionality to configure the serial output source for SGPIOs,
allowing selection between CSR, parallel input, or serial input sources.
Add a helper function to set the source and update initialization routines
to ensure proper configuration during setup.
Before the LTPI link is established, the parallel input values are
invalid. Therefore, setting the serial output via the CSR ensures that the
default value of the serial output is not undefined. Once the LTPI link is
established, the values become valid, and we can safely switch the source
back to parallel input.

Change-Id: I1ee96b2665a24f02dd2420dca8ddcd9330cb245a
Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>