dts: ast2700-ibex: update eDAF flash mode command definitions
Update the SPI0 flash mode memory base and command-write values in
the device tree file. Adjust preprocessor directives to clarify the
DDR and flash mode configurations.

These changes ensure accuracy in the memory mapping and command
definitions for SPI0 flash mode.
The DDR mode does not use the command to access the memory.

Change-Id: I0aa8c24fa7247151f62de5337705d0c0ed26dbb9
Signed-off-by: Kevin Chen <kevin_chen@aspeedtech.com>