arm: mach-aspeed: set rgmii pin to gpio
The pin default is GPIO mode, so when running RGMII calibration,
it does not send the test patterns to RGMII pin.
But, when rebooting from kernel, the pin may be configured to RGMII pin,
so it will send the test patterns to RGMII to PHY.
The test patterns is random data, it may cause some network devices
failed.

Signed-off-by: Jacky Chou <jacky_chou@aspeedtech.com>
Change-Id: Icadd0c887d13a30ddf203f1722034b9ab35d8e5c